requirements.txt
langchain>=0.0.174
deeplake>=3.5.2
python-dotenv>=1.0.0
flask>=2.3.2
